+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Unreleased
+
+## Added
+
+- added more tests and code coverage
+
+## Changed
+
+- adapted to work with recent API changes in deltachat-rpc-client-go v0.16.1-0.20230413050235-ac4cbf9913e8
+- `BotCli.Start()` now returns an error instead of calling `os.Exit(1)`
+
 ## v0.3.0
 
 - add `qr` subcommand